Login at the Start and new Leaderboard Button
The Login was moved to be enabled at the Start of the Game, at the Start Button.
The Old Options Button was changed to Leaderboard Button, and removing any associated events.
Old Leaderboard Button was removed.

NB: we did add minor changes to the Logo and the Branding of the App As well